feat(file): introduce target states and extract file contract
