+++ src/File/File.php
@@ -2,12 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Source;
+namespace Haspadar\Piqule\File;
 
 interface File
 {
-    public function exists(): bool;
-
     public function hash(): string;
 
     public function contents(): string;
