+++ src/Step/Step.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Haspadar\Piqule\Step;
 
-use Haspadar\Piqule\StepResult\StepResult;
+use Haspadar\Piqule\Step\StepResult\StepResult;
 use Haspadar\Piqule\Target\TargetFile;
 
 interface Step
